Initial brainstorms from an Etherpad
Moved over our original discussion to an actual repository, which
may (or may not), make it easier to edit... at least, easier to see what
has changed.

While this may be overkill, we would like to collect the notes we take
on the subjects we cover, and place them in this repository.